requirements.txt
langchain
langgraph
langchain-google-genai
langchain-openai
langchain-chroma
chromadb
python-dotenv
tiktoken
langchain-text-splitters
langchain-community
requests
python-dotenv
bert_score
rouge

